refactor: reorganize package structure for CustomToastState and DatePickerBottomSheet
