+++ wallet_presentation/src/commonMain/kotlin/net/thechance/mena/wallet/presentation/model/CustomToastState.kt
@@ -1,4 +1,4 @@
-package net.thechance.mena.wallet.presentation.base
+package net.thechance.mena.wallet.presentation.model
 
 import org.jetbrains.compose.resources.StringResource
